Add typed context data with generic derivation
Introduce a JSON-like `ContextData` ADT to replace the current
`Map[String, String]`, and a `ToContextData` type class to convert
things into `ContextData`. Include a few built-in implicits and generic
derivation for case classes.

The generic derivation supports the `hidden` and `secret` annotations,
so I would count this as a fix for #68.